+++ .github/workflows/assets/detekt.yml
