--- app/src/main/java/com/wndenis/snipsnap/CalendarAdapter.kt
@@ -1,4 +1,0 @@
-package com.wndenis.snipsnap
-
-data class CalendarAdapter(val name: String, val sections: List<CalendarSection>) {
-}
