+++ app/src/main/java/com/wndenis/snipsnap/MenuActivity.kt
@@ -33,6 +33,7 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.sp
 import androidx.core.content.ContextCompat.startActivity
+import com.wndenis.snipsnap.data.CalendarSection
 import java.time.LocalDateTime
