+++ detekt.bat
